app/requirements.txt
pandas
doubutsushogi
streamlit-doubutsushogi
